Combine TaskReader and TaskWriter to Storage
